move apoc and db to namespaces folder
