+++ src/namespaces/apoc/cypher/run-first-column.ts
@@ -17,12 +17,12 @@
  * limitations under the License.
  */
 
-import type { Clause } from "../../clauses/Clause";
-import type { Variable } from "../../references/Variable";
-import type { CypherEnvironment } from "../../Environment";
-import { MapExpr } from "../../expressions/map/MapExpr";
-import { CypherFunction } from "../../expressions/functions/CypherFunctions";
-import type { Expr } from "../../types";
+import type { Clause } from "../../../clauses/Clause";
+import type { Variable } from "../../../references/Variable";
+import type { CypherEnvironment } from "../../../Environment";
+import { MapExpr } from "../../../expressions/map/MapExpr";
+import { CypherFunction } from "../../../expressions/functions/CypherFunctions";
+import type { Expr } from "../../../types";
 
 /**
  * @group Cypher Functions
