+++ src/namespaces/db/fulltext.ts
@@ -18,10 +18,10 @@
  */
 
 import type { Literal, Param, Variable } from "../../Cypher";
+import { CypherProcedure } from "../../procedures/CypherProcedure";
 import type { Expr } from "../../types";
 import type { InputArgument } from "../../utils/normalize-variable";
 import { normalizeVariable, normalizeMap } from "../../utils/normalize-variable";
-import { CypherProcedure } from "../CypherProcedure";
 
 type FulltextPhrase = string | Literal<string> | Param | Variable;
